examples: SDF_example_001: decouple I/O from actors
Run separate threads which handle all input and all output respectively.
This gives us deterministic input and output ordering which is useful
for tests.

By having the input in a separate thread it's also easier to detect if
we have reached EOF. The input process will try to read all the needed
input to run a schedule once, and can exit if it was unable to before
putting the data into the buffers to the rest of the actors.

The main thread additionally waits on only the input thread, and if that
exits it kills the rest of them before exiting.