examples: better fairness in pthread fifos
In case we ever will have multiple readers and writers to the same fifo,
the current implementation releases the lock before signalling the
notempty/notfull condition variables. This means that a potential third
reader/writer could easily snatch the lock before the already waiting
ones. By signalling the condition variable before releasing the mutex,
the waiting readers/writers at least have an opportunity to take the
mutex before a new reader/writer, even if this is still not guarranteed.