SDFSchedule: do not create a new name for delay edges
Just use the input one. This should make the generated code a bit easier
to read. It should also be more efficient in the compilation steps since
we're comparing less strings.

Note that this also makes an alias for the input edge since we always
lookup the alias list in ForSyDeIRToProceduralIR.